First commit. Unintentional redundancy.
First commit. There are at least 4 copies of the Spanish translation
here. The good news is that I'm sure at least 2 of them are just old
copies/iterations. If we're lucky, it should be possible to throw out 2
or even 3 of the copies very quickly with some basic
folder/date/contents comparisons.